refs #1 A new block in creme_config to display standard relation types.

+++ creme/creme_config/views/relation_type.py
@@ -20,7 +20,7 @@
 
 from logging import debug #
 
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseRedirect, HttpResponse, Http404
 from django.shortcuts import render_to_response, get_object_or_404
 from django.template import RequestContext
 from django.contrib.contenttypes.models import ContentType
@@ -56,6 +56,9 @@ def add(request):
 def edit(request, relation_type_id):
     relation_type = get_object_or_404(RelationType, pk=relation_type_id)
 
+    if not relation_type.is_custom:
+        raise Http404("Can't edit a standard RelationType") #TODO: 403 instead ?
+
     if request.POST:
         form = RelationTypeEditForm(relation_type, request.POST)
 
@@ -76,6 +79,9 @@ def delete(request):
         @Permissions : Admin to creme_config app
     """
     relation_type = get_object_or_404(RelationType, pk=request.POST.get('id'))
+
+    if not relation_type.is_custom:
+        raise Http404("Can't delete a standard RelationType") #TODO: 403 instead ?
 
     relation_type.delete()
